# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api version 1",
        "description": "Laravel project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/catalog": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "catalog list",
                "operationId": "6d5675517445efda997387d5d2f0e2ce",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "news limit"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "параметр сортировки (price,date,stock)"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "параметр сортировки (asc,desc)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Создание товара",
                "description": "Создание нового товара.",
                "operationId": "fa88699c275219d547fdadb472ccba37",
                "requestBody": {
                    "description": "Данные для создания товара",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "title": {
                                        "description": "Название товара",
                                        "type": "string",
                                        "maxLength": 100,
                                        "minLength": 2,
                                        "example": "Новое название товара"
                                    },
                                    "description": {
                                        "description": "Описание товара",
                                        "type": "string",
                                        "minLength": 5,
                                        "example": "Подробное описание товара"
                                    },
                                    "preview_picture": {
                                        "description": "Картинка превью (до 10MB, jpeg,png,jpg,gif,svg)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[]": {
                                        "description": "Дополнительные изображения (максимум 10 файлов)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "price_opt": {
                                        "description": "Оптовая цена (число, минимум 0)",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 1500.5
                                    },
                                    "price_special": {
                                        "description": "Специальная цена (число, минимум 0)",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 1200
                                    },
                                    "price_rozn": {
                                        "description": "Розничная цена (число, минимум 0)",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 2000
                                    },
                                    "stock_ivanovo": {
                                        "description": "Количество на складе Иваново (целое число, минимум 0)",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 25
                                    },
                                    "stock_moskov": {
                                        "description": "Количество на складе Москва (целое число, минимум 0)",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 15
                                    },
                                    "stock_krasnodar": {
                                        "description": "Количество на складе Краснодар (целое число, минимум 0)",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "товар создан"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Товар не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}